updates to account for changes to loan amount columns

+++ data-browser/src/main/scala/hmda/dataBrowser/models/ModifiedLarEntity.scala
@@ -24,7 +24,7 @@ case class ModifiedLarEntity(
     reverseMortgage: Int,
     lineOfCredits: Int,
     businessOrCommercial: Int,
-    loanAmount: BigInt,
+    loanAmount: Double,
     loanValueRatio: String,
     interestRate: String,
     rateSpread: String,
